moved the DayAndNightSettingsTab class to its own file
